1. Зроблено обробку черги запитів відповідно до циклічності. 2. випралено декілька промилок із розммірностями, які виявилися на реальних даних.
git-svn-id: file:///tmp/qt4/lib@155 76bea514-bf2f-4ad3-ac3e-82dc02406d2c

+++ modbus_x/RxModbus.h
@@ -61,7 +61,7 @@ private:
 
     // Список запитів
     QVector<QByteArray> query_list;
-    QVector<qint8> query_read,local_read;
+    QVector<qint16> query_read,local_read;
 
 #ifdef ASYNC
     QQueue<QByteArray> query_queue; // черга на відправку даних в контролер
